refactor: update auto close delay type and enhance door module logic

+++ src/main/kotlin/org/xodium/vanillaplus/Config.kt
@@ -288,7 +288,7 @@ object Config {
         /**
          * The delay (in seconds) before automatic closure.
          */
-        var AUTO_CLOSE_DELAY: Int = 6
+        var AUTO_CLOSE_DELAY: Long = 6L
     }
 
     /**
